Build and install adrt when building documentation.
This will let us use the package in live examples and include the plot
output in the documentation directly. With this change autodoc can
pull documentation from the live docstrings and we will have NumPy as
well so we can disable autodoc's mock imports and remove the source
code from our path.