Add diagnostic output for Windows build steps.
Like the outputs for Linux, this makes it easier to keep an eye on
what symbols we're exporting and what other libraries we're loading.
This step is optional like the Linux diagnostics to avoid failing the
build if something changes in how we need to find and invoke the
tools.