Add caching to GitHub actions runs.
This stores downloads from PyPI in GitHub's cache. We really want a
rolling cache that gets continually evicted since some of our
dependencies may be updated to later versions. We get something like
this by tagging each cache entry with the unique workflow run ID, but
accept restoring cache entries without it.

The restore keys depend on the Python version. This does two things:
first, native dependencies often depend on this (like NumPy) so we get
more usable cache entries on restore; and second it limits the
lifetime of the cache to that of a particular Python release. This
ensures that regardless cache usage all cache entries expire at the
next Python release. That way we don't have ancient Pip cache contents
that can get carried forward indefinitely.

Building release artifacts doesn't use the cache to make sure we get
everything fresh.

The new cleanup step removes any copies of our package from the cache.
We hide the error code at the shell level so we don't get warnings
from the job if Pip fails at this step. In particular, Pip produces an
error if it finds no matches to remove; this is fine for us so we hide
the error. Using GitHub's "continue-on-error" produces a visible
warning when the step fails.