Explicitly specify Python versions in workflows.
Unspecified versions may cause issues after a new Python release until
both NumPy and SciPy have pre-built releases available. NumPy we could
work around by spending the time to build it ourselves, but SciPy is
more complicated. Tests and releases are either run on the full range
of supported versions, or on a fixed version at one of the two ends of
the range. When it's time to support a new Python release the lists of
versions will have to be updated to add the new version and remove
newly unsupported releases. Fixed version numbers in workflows (either
oldest or latest) should be updated to keep them consistent with the
list of supported releases.