Split adrt implementation into separate functions.
This begins the process of separating out the core ADRT loop from the
rest of the implementation. We also keep Python-dependent parts out of
the adrt implementation itself (mainly lift the buffer allocation into
the Python glue code). Some shape checking and computing functions are
moved into normal non-template functions as these are not performance
critical.