Add additional explicit casts and range checks to NumPy conversions.
When converting values to and from NumPy shapes make sure we don't
overflow and use explicit casts to silence compiler warnings about
conversion between signed and unsigned types. On the input side, we
don't overflow because size_t is large enough for the full buffer's
size. On the output size, explicitly check the range and fail if
needed. We use an error similar to the one NumPy signals in the same
case.