Simplify basic Optional use to manipulate more values in place.
The adjustments make it possible to fill in values inside a
constructed Optional value. This relies on the default initialization
which works fine for the simple types we use. Also, the methods and
operators used to extract the values now align with C++17's
std::optional. Constructing the values is different, but consuming
them should work the same way as std::optional.