Remove implicit barriers from inner loop in iadrt.
This relies on OpenMP static loop schedule to ensure that previous
values are written to the previous row by the same thread that will
read them next. Comments added to the code to explain this since it is
more fragile than other functions. This reduces the overhead of
barriers and makes more of a difference for small inputs.