Reorder accesses in iadrt core loop.
Assign each thread to iterate along the full row before proceeding to
other columns. This simplifies the OpenMP accesses so that we don't
critically depend on a static loop schedule and still avoid the
barriers which were present previously. This also produces a larger
speedup given more friendly memory accesses in the current layout.