Add support for NumPy integer types to num_iters wrapper.
If the argument is a Python integer, a NumPy integer, or another class
registered with the numbers.Integral ABC, coerce it to a Python
integer and call the C++ function. Otherwise raise a TypeError with
the type of the argument.

We do the first checks against the explicit int type and the
np.integer class to avoid the more expensive check against the ABC's
registry. In a quick test, this added support doubles the time taken
in the wrapper, but a check against the ABC doubles it again so we
avoid that for more likely argument types. The wrapper is still quite
fast so the doubling we're incurring here shouldn't be a problem.