Coerce to integer using operator.index in num_iters.
Instead of checking the type of the argument we can use operator.index
to losslessly convert to integers for appropriate types. This could be
done using the C API PyNumber_Index, but for a small slowdown we can
do it in Python like we do for array type conversions and avoid more
refcount management in native code.