Use oldest-supported-numpy to determine NumPy version for wheels.
This should better handle the case where a wheel is built
automatically while the package is installed on a platform where we do
not distribute pre-built wheels.

For that, we use a bit of a hack: we ensure the wheel ABI tag supports
only the current version of Python and later. That we we've used the
oldest version of NumPy for those Pythons when building the wheel.

Otherwise, users might built a wheel with a later version of Python
and install on an earlier version, which should work except if there
is a NumPy version mismatch. For cases where the wheel is installed
immediately this is invisible, but in other cases users have to build
with the same version of Python they are targeting (or override the
option).