Simplify usage of unpack_tuple and warn if return value is unused.
Allow only specifying one template size parameter when both are
equal (require always exactly the same number of arguments). This is
the only case we use at the moment so it simplifies our current use.
Also mark the function so that GCC and Clang will warn if the result
is not checked.