Manually specify manylinux version to use for releases.
Fix the manylinux container version we use for builds. We want
primarily the most recent GCC version available which is currently in
manylinux2014. We specify containers for all the platforms we build
Linux wheels for (currently x64 only) and should add other pins here
if we build on additional Linux architectures in the future.