Manually unroll shape computation and array stride access loops.
This significantly reduces our dependence on aggressive compiler loop
unrolling. With this change the performance gap between GCC O2 and O3
is nearly closed. These two unrolled loops are used in each array
access and it is critical that they be inlined and unrolled so that
portions which are loop-invariant can be moved out of the
loops (stride computations, etc.). Modern GCC does the unrolling we
need at O3, but not O2 so we unroll manually with template parameters.