Change exception type for bad argument to new_array.
This case shouldn't ever happen and would indicate a bug rather than
user error. Make this clearer in the message and change exception type
to AssertionError (as though this were a Python assert) instead of
RuntimeError. We still check this to guard against memory corruption
and because this function isn't performance critical.