Remove use of default initialization in Optional.
This seems to technically lead to undefined behavior if the copy
constructor is invoked when the Optional's value contains an
indeterminate value. Avoid the problem the way std::optional does
using a union. Update all users to new constructors. Switch to
parentheses in initializer list to avoid errors with std::array brace
initialization in old GCC.