Make shape_product commutative with zeros after overflow.
We don't anticipate zero shapes so this isn't critical, but does
remove a possible surprising behavior. Before this, if a shape started
with zero, even if later values would cause overflow in the product
the result would correctly be zero. However an overflow before zero
would not be reset to zero. There are no other values which could
bring values back in range (no subtleties with negatives or similar)
so we just check for zeros. We don't short circuit here since we
anticipate always multiplying all values together and this seems to
keep GCC inlining this function.