Add explicit cap on size of input arrays.
This change should not add any extra restriction on inputs that could
actually exist, just makes on particular limit more explicit. In the
single-step functions, we perform a left shift to compute some
divisors and we need to make sure that these do not go out of range.
Inputs that large cannot actually exist because we currently require
square images.