Add output shape computing functions for single-step routines.
These just return the same array as their input since the output shape
is the same as the input, but it makes the single-step flow and usage
match the full routines more closely. That is, we have functions that
produce the required shapes for all buffers that need to be allocated.