Update test workflow for each push and pull request.
Run the code base on four types of tests: linters, Python test suite
with OpenMP, Python test suite with sanitizers, and a C++ test suite
with sanitizers. This allows us to exercise the C++ utility functions
directly and gives us coverage from the sanitizers in addition to our
own assertions.