Use Catch2 check macros for most test conditions.
Many of the conditions we check aren't mandatory for the rest of the
test to proceed. Use the check macro to mark the test as failed, but
still evaluated the remaining conditions. One exception is checking
whether an Optional has a value, these need to cancel the rest of the
test and still use the require macro. Also make use of the _false
family of macros to remove negated test conditions.