Use unsigned types in arithmetic and comparisons.
In particular, use size_t constants in arithmetic on array indices to
ensure that the computation is not subject to conversions. The basic
rules followed here are that size_t arithmetic should be done with all
values as size_t, but comparisons against a constant may use unsigned
int literals (but should not use signed int literals, so add the u
suffix). Direct assignments of a small constant to a size_t variable
may still use regular int literals.