Add global include for size_t in Python bindings.
We provide one for everything inside our adrt namespace (in
adrt_cdefs_common.hpp), but this doesn't cover our extern "C"
functions which we put directly in the global namespace. Python also
includes a global size_t definition, but we don't want to rely on that
so we provide our own as well.