Add additional assertions around memory allocations.
One of these is converted from an always-present runtime check to a
regular test-only assertion. The condition it was guarding is one
which we expect to always be true and which would be no worse on
failure than other guards we've left as assertions. Add others
ensuring that we don't allocate empty buffers or arrays with zero
dimensions.