Begin separating content and index float types and checking sizes.
Add separate template parameter for float types used for index
computations to separate this choice from the choice of the type of
the array contents. Also add another template function for verifying
that input arrays are inside the size limits that we can correctly
process as float. That limit still needs to be checked.