Set pytest configuration options in pyproject.toml.
This requires a pytest>=6 so add that constraint to our test
requirements. For options, we disable stdout capturing to ensure that
any native assertions or sanitizer messages are printed before the
program exits. Pytest capturing can't handle this since control never
returns to the interpreter. We also enable importlib as suggested by
the pytest docs.