Remove OpenMP SIMD directives.
The documentation says that without a manual "aligned" clause these
assume implementation-defined default alignments for the target
platform's SIMD instructions. This might work out for us in most cases
but we aren't in a position to guarantee anything for our allocated
memory. Even specifying an alignment of 1 might lead the compiler to
be too pessimistic. Best just to remove the directives and let the
auto-vectorizer decide what to do on its own.