Move detailed build flags into release workflow.
Specify only basic build flags in the configuration file. In
particular, optimization and visibility flags, but not hardening. The
specific flags we want for release builds may (in the future) depend
on architecture. This change paves the way for specifying some of
these flags as part of the workflow matrix like we do in our test
workflow.