Check that skipped dimensions are trivial in new_array.
Make sure that any dimensions we aren't reflecting in new_array are
trivial dimensions which would not affect the overall allocation
side (i.e. dimensions of size 1). These are assumed to be added by us
for missing optional batch dimensions and we want to skip them in the
output we return to the user.