Use C++ algorithms for array shape assertions.
This avoids needing to add our own conditional compilation blocks in
addition to those used internally by the assert macro. In general our
use is pretty straightforward. However, for the bounds check in
array_access we create a temporary array from the parameter pack
arguments. This is fine in our current use since the temporary array
will live for the full-expression of the assert macro which is long
enough for our evaluation.