Add PyPI upload step to release workflow.
The final two steps upload the package to PyPI and then make the
GitHub release visible. The GitHub release is a separate job so we can
minimize the number of steps which have write access to the repository
and so that those steps don't download anything to the runner other
than our own artifacts.