Add automated test for NumPy build version.
The version of NumPy used to build the package is chosen by
oldest-supported-numpy. This is the value that should be used for the
NPY_NO_DEPRECATED_API macro. Add an automated check that we are using
the proper value so that we don't forget to increase it if we raise
our minimum supported Python version.

This check is somewhat complicated because the only thing we can do
with requirement markers is evaluate them. So we have to try all
platform values that appear. These are sourced from the test suite
used in oldest-supported-numpy and from the Python documentation for
the platform module. Ideally, we should periodically check these and
update the list with any that are added or that are in use.