Add more assertions for float indices in interpolation routine.
These check that the float index values are finite. This is relevant
for si in particular since it has no other range checks before we cast
it to a size_t which would be undefined behavior for out-of range
values. For hi we do have a bounds check so we just check finiteness
for good measure.