Use only-binary for documentation requirements.
Some of these dependencies would be hard to build from source and we
would rather fail the docs build quickly rather than waste time on the
shared runners trying to build them. Generally, making sure we use a
supported, but not too new, version of Python should be enough to
avoid issues.