Use NumPy 1.25 during the build.
The latest version of NumPy includes a feature to use a backwards
compatible API by setting a preprocessor macro before including. In
principle this should replace our use of oldest-supported-numpy and
allow us to build using NumPy directly. For now, we exclude 2.0 until
the picture around the next major version is clearer.