Remove explicit setting of NumPy API target.
Using the default value will simplify our packaging. In particular,
conda-forge should correctly handle this lower limit for us. If we
override the value, we would then need to add additional handling for
the Conda packages. The runtime limit should still be updated
occasionally in line with SPEC 0 for PyPI builds.